+++ src/main/java/com/tristankechlo/toolleveling/config/values/IConfigValue.java
@@ -1,0 +1,17 @@
+package com.tristankechlo.toolleveling.config.values;
+
+import com.google.gson.JsonObject;
+
+public interface IConfigValue<T> {
+
+	String getIdentifier();
+
+	void setToDefault();
+
+	T getValue();
+	
+	void serialize(JsonObject jsonObject);
+
+	void deserialize(JsonObject jsonObject);
+
+}